+++ commands/utility.js
@@ -591,13 +591,14 @@ module.exports = [
 		constructor() {
 			super({
 				name: "userinfo",
-				description: "Get info about a user",
+				description: "Get info about a user. You can also provide any user ID to get info even if outside this server",
 				aliases: ["member", "memberinfo", "user"],
 				args: [
 					{
 						infiniteArgs: true,
 						optional: true,
-						type: "member"
+						type: "member",
+						allowRaw: true
 					}
 				],
 				cooldown: {
@@ -614,53 +615,70 @@ module.exports = [
 		}
 		
 		async run(bot, message, args, flags) {
-			const member = args[0] || message.member;
-			let userRoles = member.roles.array();
-			userRoles.splice(userRoles.findIndex(role => role.id == message.guild.id), 1);
-			userRoles = userRoles.map(role => role.name);
-			
-			const guildMembers = message.guild.large ? await fetchMembers(message) : message.guild.members,
-				guildMemArray = guildMembers.array();
-			guildMemArray.sort((a,b) => a.joinedTimestamp - b.joinedTimestamp);
-			
-			const joinPos = guildMemArray.findIndex(mem => mem.joinedTimestamp == member.joinedTimestamp), nearbyMems = [];
-			for (let i = joinPos - 2; i < joinPos + 3; i++) {
-				if (i < 0 || i >= message.guild.memberCount) continue;
-				nearbyMems.push(i == joinPos ? `**${guildMemArray[i].user.username}**` : guildMemArray[i].user.username);
-			}
-			
-			const memPresence = member.presence;
-			let userPresence;
-			if (memPresence.status == "online") {
-				userPresence = "Online";
-			} else if (memPresence.status == "idle") {
-				userPresence = "Idle";
-			} else if (memPresence.status == "dnd") {
-				userPresence = "Do Not Disturb";
+			const userEmbed = new RichEmbed();
+			let user, member;
+			if (typeof args[0] == "string") {
+				let cmdErr = true;
+				if (args[0].length >= 17 && args[0].length < 19 && !isNaN(args[0])) {
+					await bot.fetchUser(args[0])
+						.then(fetchedUser => {user = fetchedUser; cmdErr = false})
+						.catch(() => {});
+				}
+				if (cmdErr) return {cmdWarn: "No users found matching `" + args[0] + "`"};
 			} else {
-				userPresence = "Offline";
+				member = args[0] || message.member;
+				user = member.user;
 			}
-			if (memPresence.game) userPresence += ` (playing ${memPresence.game.name})`;
-			
-			const createdDate = new Date(member.user.createdTimestamp),
-				joinedDate = new Date(member.joinedTimestamp);
-			
-			const userEmbed = new RichEmbed()
-				.setTitle(`User Info - ${member.user.tag}`)
-				.setThumbnail(member.user.avatarURL || `https://cdn.discordapp.com/embed/avatars/${member.user.discriminator % 5}.png`)
-				.setFooter(`ID: ${member.id}`)
-				.addField("Account created at", `${createdDate.toUTCString()} (${getDuration(createdDate)})`)
-				.addField("Joined this server at", `${joinedDate.toUTCString()} (${getDuration(joinedDate)})`)
-				.addField("Status", userPresence, true)
-				.addField("Nickname", member.nickname || "None", true)
-				.addField("Bot user", member.user.bot ? "Yes" : "No", true)
-				.addField("Member #", joinPos + 1, true)
-				.addField("Join order", `${nearbyMems.join(" > ")}`)
-				.addField(`Roles - ${userRoles.length}`, userRoles.length == 0 ? "None" : userRoles.join(", "));
-			
-			if (member.displayColor != 0 || (member.colorRole && member.colorRole.color == 0)) {
-				userEmbed.setColor(member.displayColor);
+			const rawPresence = (member && member.presence) || user.presence;
+			let presence;
+			if (rawPresence.status == "online") {
+				presence = "Online";
+			} else if (rawPresence.status == "idle") {
+				presence = "Idle";
+			} else if (rawPresence.status == "dnd") {
+				presence = "Do Not Disturb";
+			} else {
+				presence = "Offline";
 			}
+			if (rawPresence.game) presence += ` (playing ${rawPresence.game.name})`;
+				
+			const createdDate = new Date(user.createdTimestamp);
+			userEmbed.setTitle(`User Info - ${user.tag}`)
+				.setThumbnail(user.avatarURL || `https://cdn.discordapp.com/embed/avatars/${user.discriminator % 5}.png`)
+				.setFooter(`ID: ${user.id}`)
+				.addField("Account created at", `${createdDate.toUTCString()} (${getDuration(createdDate)})`);
+
+			if (member) {
+				const joinedDate = new Date(member.joinedTimestamp);
+				userEmbed.addField("Joined this server at", `${joinedDate.toUTCString()} (${getDuration(joinedDate)})`);
+			}
+			userEmbed.addField("Status", presence, true)
+				.addField("Bot user", user.bot ? "Yes" : "No", true);
+			if (member) {
+				let userRoles = member.roles.array();
+				userRoles.splice(userRoles.findIndex(role => role.id == message.guild.id), 1);
+				userRoles = userRoles.map(role => role.name);
+				
+				const guildMembers = message.guild.large ? await fetchMembers(message) : message.guild.members,
+					guildMemArray = guildMembers.array();
+				guildMemArray.sort((a,b) => a.joinedTimestamp - b.joinedTimestamp);
+				
+				const joinPos = guildMemArray.findIndex(mem => mem.joinedTimestamp == member.joinedTimestamp), nearbyMems = [];
+				for (let i = joinPos - 2; i < joinPos + 3; i++) {
+					if (i < 0 || i >= message.guild.memberCount) continue;
+					nearbyMems.push(i == joinPos ? `**${guildMemArray[i].user.username}**` : guildMemArray[i].user.username);
+				}
+
+				userEmbed.addField("Nickname", member.nickname || "None", true)
+					.addField("Member #", joinPos + 1, true)
+					.addField("Join order", `${nearbyMems.join(" > ")}`)
+					.addField(`Roles - ${userRoles.length}`, userRoles.length == 0 ? "None" : userRoles.join(", "));
+
+				if (member.displayColor != 0 || (member.colorRole && member.colorRole.color == 0)) {
+					userEmbed.setColor(member.displayColor);
+				}
+			}
+			
 			message.channel.send(userEmbed);
 		}
 	}