+++ structures/flag.js
@@ -9,7 +9,7 @@ class Flag {
 
 class FlagArgument {
 	constructor(props) {
-		this.errMsg = props.errMsg || null;
+		this.errorMsg = props.errorMsg || null;
 		this.num = props.num || 1;
 		this.optional = props.optional || false;
 		this.type = props.type;